+++ src/components/games/poker/PokerTable.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const PokerTable = () => {
+  return <div>PokerTable</div>;
+};
