+++ src/components/games/rummy/RummyTable.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const RummyTable = () => {
+  return <div>RummyTable</div>;
+};
