+++ src/components/games/speed/SpeedTable.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const SpeedTable = () => {
+  return <div>SpeedTable</div>;
+};
