+++ src/components/games/texasPoker/TexasPokerTable.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const TexasPokerTable = () => {
+  return <div>TexasPokerTable</div>;
+};
