--- src/components/model/player/sample.ts
@@ -1,1 +1,0 @@
-export const message = "ディレクトリ構成をコミットに反映するため、便宜上で作ったファイルです"
