--- src/components/model/player/war/TestHand.tsx
@@ -1,89 +1,0 @@
-import { Card } from "@/models/card/card";
-import React from "react";
-import { CardComponent } from "../../card/CardComponent";
-
-export const TestHand = () => {
-  return (
-    <>
-      <div>
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-
-      <div className="absolute top-0 left-[10px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[20px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[30px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[40px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[50px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[60px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[70px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[80px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[90px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[100px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[110px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[120px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[130px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[140px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[150px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[160px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[170px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[180px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[190px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[200px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[210px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[220px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[230px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[240px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-      <div className="absolute top-0 left-[250px]">
-        <CardComponent card={new Card("10", "C", false, "/cards/10-C.png")} />
-      </div>
-    </>
-  );
-};
