+++ src/components/page/SpeedPage.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { GamePageLayout } from "../common/ui/layout/GamePageLayout";
+import { SpeedTable } from "../games/speed/SpeedTable";
+
+export const SpeedPage = () => {
+  return (
+    <GamePageLayout>
+      <SpeedTable />
+    </GamePageLayout>
+  );
+};
