--- src/components/ui/layout/HeaderLayout.tsx
@@ -1,15 +1,0 @@
-import React, { FC, ReactNode } from "react";
-import { Header } from "../organisms/Header";
-
-type Props = {
-  children: ReactNode;
-};
-export const HeaderLayout: FC<Props> = (props) => {
-  const { children } = props;
-  return (
-    <>
-      <Header />
-      {children}
-    </>
-  );
-};
