+++ src/hooks/games/poker/usePokerState.ts
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const usePokerState = () => {
+  return {};
+};
