+++ src/hooks/games/rummy/useRummyState.ts
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const useRummyState = () => {
+  return {};
+};
