+++ src/hooks/games/spped/useSpeedState.ts
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const useSpeedState = () => {
+  return {};
+};
