+++ src/hooks/games/texasPoker/useTexasPokerState.ts
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const useTexasPokerState = () => {
+  return {};
+};
