+++ src/hooks/games/war/useWarState.ts
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const useWarState = () => {
+  return {};
+};
